+++ pytsite/odm_auth/_api.py
@@ -1,70 +1,48 @@
 """PytSite ODM Auth API Functions.
 """
 from typing import Iterable as _Iterable
-from pytsite import permissions as _permission, auth as _auth, odm as _odm
-from . import _model
+from pytsite import auth as _auth, odm as _odm, permissions as _permissions
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
-def check_permissions(action: str, model: str, ids: _Iterable = None, user: _auth.model.AbstractUser = None) -> bool:
+def check_permission(perm: str, model: str, ids: _Iterable = None, user: _auth.model.AbstractUser = None) -> bool:
     """Check current user's permissions to operate with entity(es).
     """
+    perm = 'pytsite.odm_auth.{}.{}'.format(perm, model)
+
     # Get current user
     if not user:
         user = _auth.get_current_user()
 
-    # Check ids type
-    if ids and type(ids) not in (list, tuple):
+    # Check for permission existence and whether user has it
+    if not (_permissions.is_permission_defined(perm) and user.has_permission(perm)):
+        return False
+
+    if isinstance(ids, str):
         ids = (ids,)
 
-    if action == 'create':
-        create_perm_name = 'pytsite.odm_auth.create.' + model
-        if _permission.is_permission_defined(create_perm_name) and user.has_permission(create_perm_name):
-            return True
-    else:
-        # If 'global' permission was not defined
-        global_perm_name = 'pytsite.odm_auth.' + action + '.' + model
-        if not _permission.is_permission_defined(global_perm_name):
-            return False
+    # Check user's personal permission
+    if perm.find('_own.') and isinstance(ids, (list, tuple)):
+        # Check each entity
+        for eid in ids:
+            entity = _odm.dispense(model, eid)
 
-        # If user has 'global' permissions for model
-        if user.has_permission(global_perm_name):
-            return True
+            # Anyone cannot do anything with non-existent entities
+            if not entity:
+                return False
 
-        # If 'personal' permission was not defined
-        personal_perm_name = 'pytsite.odm_auth.' + action + '_own.' + model
-        if not _permission.is_permission_defined(personal_perm_name):
-            return False
+            # Searching for author of the entity
+            for author_field in 'author', 'owner':
+                if entity.has_field(author_field):
+                    author = entity.f_get(author_field)
 
-        # Else check user's personal permission
-        if user.has_permission(personal_perm_name):
-            if ids:
-                # Check each entity
-                for eid in ids:
-                    entity = _odm.dispense(model, eid)
+                    if not isinstance(author, _auth.model.AbstractUser) and author is not None:
+                        raise RuntimeError('Entity must return user object or None here.')
 
-                    # Anyone cannot do anything with non-existent entities
-                    if not entity:
+                    if author != user:
                         return False
 
-                    # Searching for author of the entity
-                    for author_field in 'author', 'owner':
-                        if entity.has_field(author_field):
-                            author = entity.f_get(author_field)
-
-                            if not isinstance(author, _auth.model.AbstractUser) and author is not None:
-                                raise RuntimeError('Entity must return user object or None here.')
-
-                            if not user.is_anonymous and author == user:
-                                return True
-
-                            # Entity belongs to nobody
-                            elif author is None:
-                                return True
-            else:
-                return True
-
-    return False
+    return True