+++ pytsite/odm_auth/_eh.py
@@ -51,14 +51,14 @@ def odm_entity_pre_save(entity: _model.AuthorizableEntity):
     if c_user.is_system or c_user.is_admin:
         return
 
-    # Check current user's permissions to MODIFY entities
-    if entity.is_new and not entity.check_permissions('create'):
+    # Check current user's permissions to CREATE entities
+    if entity.is_new and not entity.odm_auth_check_permission('create'):
         _logger.info('Current user login: {}'.format(_auth.get_current_user().login))
         raise _errors.ForbidCreation("Insufficient permissions to create entities of model '{}'.".
                                      format(entity.model))
 
     # Check current user's permissions to MODIFY entities
-    elif not entity.is_new and not entity.check_permissions('modify'):
+    if not entity.is_new and not (entity.odm_auth_check_permission('modify') or entity.odm_auth_check_permission('modify_own')):
         _logger.info('Current user login: {}'.format(_auth.get_current_user().login))
         raise _errors.ForbidModification("Insufficient permissions to modify entity '{}:{}'.".
                                          format(entity.model, entity.id))
@@ -78,7 +78,7 @@ def odm_entity_pre_delete(entity: _model.AuthorizableEntity):
         return
 
     # Check current user's permissions to DELETE entities
-    if not entity.check_permissions('delete'):
+    if not (entity.odm_auth_check_permission('delete') or entity.odm_auth_check_permission('delete_own')):
         _logger.debug('Current user login: {}'.format(_auth.get_current_user().login))
         raise _errors.ForbidDeletion("Insufficient permissions to delete entity '{}:{}'.".
                                      format(entity.model, entity.id))